examples/experimental-freethreaded/pyproject.toml
[project]
name = "experimental-freethreaded"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "blazefl>=2.0.0dev7",
    "hydra-core>=1.3.2",
    "torch>=2.7.1",
    "torchvision>=0.22.1",
]

[tool.basedpyright]
typeCheckingMode = "standard"

[[tool.mypy.overrides]]
module = ["torchvision.*"]
ignore_missing_imports = true

[tool.ruff.lint]
select = [
    "E",  # pycodestyle
    "F",  # Pyflakes
    "UP",  # pyupgrade
    "B",  # flake8-bugbear
    "SIM",  # flake8-simplify
    "I",  # isort
]
ignore = []
fixable = ["ALL"]

[dependency-groups]
dev = [
    "mypy>=1.16.0",
    "types-tqdm>=4.67.0.20250516",
]
